Problem: Rust has no equivalent of JMX (Java Management Extensions)
Even though JMX is definitely rough on the edges, it allows for
great introspection and operational capabilities for all Java
systems.

Solution: develop a minimal prototype for achieving similar objectives
in Rust.

Current prototype does not allow for sending messages into the system
yet (only instrument readings) and doesn't have a way to serve or publish
this data to external consumers.