Problem: no support for external consumers
Currently, it's only possible to read instruments from within the application
itself, meaning one has to implement a network server or a published to share
instruments' data.

Solution: implement a feature-gated MQTT client that will publish instruments
to an MQTT broker.

The idea behind this is to immediately enable streaming updates for instruments
and allow routing of this data to interested parties without implementing
anything complicated.

Closes #1